The last four questions leave the header-map module
`word = token / quoted-string` gets the name its own specification gives it —
and getting there is the correction of the plan's first guess, which filed
`token_or_quoted_string` under `token`. It transcribes and cites *both*
alternatives, so that would have named it for half its grammar.

`qvalue` is small and separate for the same reason `status` is: one question,
answered as a production rather than as a list of the strings the production
generates. `field_placement` holds the three predicates that take a field
*name* and answer where it may appear — none of them reads a value, which is
exactly why they were not the header-map question.

`ext-value` joins `parameter` rather than getting a module. It is not a
separate question: it is what a parameter value may be when the name ends in
`*`, and every reader of it already reads parameters.

One integration test in `lint-http-proxy/tests` imported a moved helper through
the `lint_http` re-export — outside every `src/` sweep this batch had run.
`cargo lint` caught it; all test directories are now swept too.

headers.rs: 2344 → 346 non-test lines, 70 top-level items → 13.
Its imports are now exactly one line: `use hyper::HeaderMap;`.
