+++ lint-http-rules/src/helpers/headers.rs
@@ -2,9 +2,6 @@
 //
 // SPDX-License-Identifier: ISC
 
-use crate::helpers::list::list_members;
-use crate::helpers::quoted_string::unescape_quoted_string;
-use crate::helpers::shown::describe_char;
 use hyper::HeaderMap;
 
 /// Retrieve a header value as a string, if it exists and contains only visible ASCII.
@@ -334,171 +331,6 @@ mod concat_header_tests {
         assert_eq!(get_all_header_values(&headers, "bar"), None);
     }
 }
-/// Field names that cannot appear in a trailer section, by category.
-///
-/// This is the fixed half of the question "may this field be a trailer". The other
-/// half is [`is_nominated_by_connection`], which depends on the message.
-///
-/// **This table is a subset, and deliberately so.** The requirement is stated the
-/// other way round — a trailer field is forbidden *unless* the sender knows the
-/// field's own definition permits it, and the registry guidance says that by default
-/// no definition does. A table of names answers the opposite question, so every field
-/// nobody thought of passes. It cannot be inverted here: for a field this codebase
-/// holds no definition of (`X-Checksum`, `Grpc-Status`), only the sender knows whether
-/// a definition permits the usage, and reporting all of them would report the senders
-/// that read their own specification. What the table can hold honestly is the fields
-/// whose definitions are in the specifications this crate cites and do *not* permit
-/// trailers; `trailer_fields_valid` says so where an operator reads it.
-///
-/// The categories below are the ones §6.5.1 names, in its order. The last group is
-/// here because a connection-specific field is, by definition, needed before the
-/// content is read; `trailer` is here because a Trailer field inside a trailer
-/// section announces nothing.
-///
-/// Membership is per field *definition*, never per category: RFC 9110 puts
-/// `Authentication-Info` and `Proxy-Authentication-Info` under authentication and
-/// then permits both in trailers, so neither is here. See
-/// [`is_prohibited_trailer_field`] for where that is written down.
-///
-// cite(RFC 9110 § 6.5.1): "Many fields cannot be processed outside the header section because their evaluation is necessary prior to receiving the content, such as those that describe message framing, routing, authentication, request modifiers, response controls, or content format."
-// cite(RFC 9110 § 6.5.1): "A sender MUST NOT generate a trailer field unless the sender knows the corresponding header field name's definition permits the field to be sent in trailers."
-// cite(RFC 9110 § 16.3.2): "If the field is allowable in trailers; by default, it will not be (see Section 6.5.1)."
-pub static PROHIBITED_TRAILER_FIELDS: &[&str] = &[
-    // Message framing
-    "content-length",
-    "transfer-encoding",
-    // Routing
-    "host",
-    // `Forwarded` is routing too, and RFC 7239 defines it for the header
-    // section and nowhere else — so the sender does not know its definition
-    // permits a trailer, because it does not.
-    "forwarded",
-    // Request modifiers — controls
-    "cache-control",
-    "expect",
-    "max-forwards",
-    "pragma",
-    "range",
-    "te",
-    // Request modifiers — conditionals
-    "if-match",
-    "if-modified-since",
-    "if-none-match",
-    "if-range",
-    "if-unmodified-since",
-    // Authentication (RFC 9110 §11). `Authentication-Info` and
-    // `Proxy-Authentication-Info` are the two that §11 takes back out again, and
-    // they are not here.
-    "authorization",
-    "proxy-authenticate",
-    "proxy-authorization",
-    "www-authenticate",
-    // Response control data. These are the fields §6.5.1's "response controls"
-    // names in prose; §6.2 "Control Data" is the start line, not this.
-    "age",
-    "date",
-    "expires",
-    "location",
-    "retry-after",
-    "vary",
-    "warning",
-    // Payload processing
-    "content-encoding",
-    "content-range",
-    "content-type",
-    "trailer",
-    // Connection-specific (RFC 9110 §7.6.1). The section's own list is bullets
-    // too short to cite; the citable form is RFC 9113's parenthetical, a
-    // published reading of that list, and it names `Proxy-Connection` in it —
-    // the member this table omitted while `CONNECTION_SPECIFIC_FIELDS` below
-    // held it, which is what made the omission a reading rather than a
-    // decision (RULECITES P46). `transfer-encoding` and `te` are §7.6.1's too
-    // and sit above under the framing and request-modifier categories §6.5.1
-    // names them by.
-    // cite(RFC 9113 § 8.2.2): "This includes the Connection header field and those listed as having connection-specific semantics in Section 7.6.1 of [HTTP] (that is, Proxy-Connection, Keep-Alive, Transfer-Encoding, and Upgrade)."
-    "connection",
-    "keep-alive",
-    "proxy-connection",
-    "upgrade",
-    // The cleanest member of this list: § 6.5.1 asks whether the field's own
-    // definition permits the usage, and this field's definition answers by name.
-    // The same sentence forbids it in a response, which is
-    // `early_data_header_safe_method`'s finding — a response *trailer* is
-    // both, and is reported here.
-    // cite(RFC 8470 § 5.1): "An Early-Data header field MUST NOT be included in responses or request trailers."
-    "early-data",
-];
-
-/// Whether `name` cannot appear in a trailer section because of what it is.
-///
-/// See [`is_nominated_by_connection`] for the half that depends on the message.
-///
-/// Two fields this returns `false` for used to be in the table, on the strength of
-/// the category §6.5.1 lists them under. Both definitions say the opposite in one
-/// sentence each, and the sentence is the thing §6.5.1 asks the sender to know — so
-/// a sender whose scheme allows it is conforming, and the rule that reports this
-/// cannot tell which scheme is in use. The permission is conditional and the
-/// condition is not on the wire; a name the RFC permits at all does not belong in a
-/// table of names the RFC forbids.
-///
-// cite(RFC 9110 § 11.6.3): "Authentication-Info can be sent as a trailer field (Section 6.5) when the authentication scheme explicitly allows this."
-// cite(RFC 9110 § 11.7.3): "Proxy-Authentication-Info can be sent as a trailer field (Section 6.5) when the authentication scheme explicitly allows this."
-pub fn is_prohibited_trailer_field(name: &str) -> bool {
-    let name_l = name.trim().to_ascii_lowercase();
-    PROHIBITED_TRAILER_FIELDS.contains(&name_l.as_str())
-}
-
-/// Fields that are connection-specific whatever the message says.
-///
-/// The first six are RFC 9110 § 7.6.1's own list, plus `connection` itself, which the
-/// same section has an intermediary remove after acting on it. The last two are not
-/// § 7.6.1's: they are single-hop by their own definitions in § 11.7.1 and § 11.7.2.
-///
-/// `trailer` is deliberately not here — § 6.6.2 has it surviving the hop. It cannot be
-/// a *trailer field*, but that is § 6.5.1's business and
-/// [`PROHIBITED_TRAILER_FIELDS`] is where it says so.
-// cite(RFC 9110 § 7.6.1): "Furthermore, intermediaries SHOULD remove or replace fields that are known to require removal before forwarding, whether or not they appear as a connection-option, after applying those fields' semantics."
-static CONNECTION_SPECIFIC_FIELDS: &[&str] = &[
-    "connection",
-    "keep-alive",
-    "proxy-connection",
-    "te",
-    "transfer-encoding",
-    "upgrade",
-    // Single-hop by their own definitions, not by § 7.6.1's list.
-    "proxy-authenticate",
-    "proxy-authorization",
-];
-
-/// Whether `name` is connection-specific — either always, or because this message's
-/// `Connection` header nominated it.
-pub fn is_connection_specific_field(name: &str, connection_header_value: Option<&str>) -> bool {
-    let name_l = name.trim().to_ascii_lowercase();
-    CONNECTION_SPECIFIC_FIELDS.contains(&name_l.as_str())
-        || is_nominated_by_connection(name, connection_header_value)
-}
-
-/// Whether `name` was named as a connection-option in this message's `Connection`
-/// header, which disqualifies it from the trailer section for this message only.
-///
-/// The cited sentence says "header **or trailer** field(s)", and that is the whole
-/// reason a connection-option reaches into a trailer section at all.
-///
-/// Every caller passes a value joined by [`combined_field_value_as_written`], i.e.
-/// one `char` per octet, so the member walk is the `OWS`-trimming one: a member
-/// padded with an `obs-text` octet that renders like a space is not the
-/// `connection-option` it resembles, and no sentence lets this function pretend it
-/// is. The fold *is* licensed -- a `connection-option` is a field name, and those
-/// are case-insensitive.
-// cite(RFC 9110 § 5.1): "Field names are case-insensitive"
-pub fn is_nominated_by_connection(name: &str, connection_header_value: Option<&str>) -> bool {
-    // cite(RFC 9110 § 7.6.1): "Intermediaries MUST parse a received Connection header field before a message is forwarded and, for each connection-option in this field, remove any header or trailer field(s) from the message with the same name as the connection-option, and then remove the Connection header field itself (or replace it with the intermediary's own control options for the forwarded message)."
-    let name_l = name.trim().to_ascii_lowercase();
-    let Some(conn) = connection_header_value else {
-        return false;
-    };
-    list_members(conn).any(|tok| tok.eq_ignore_ascii_case(name_l.as_str()))
-}
 
 /// The sentence a singleton field says when a message carries more than one of
 /// its field lines.
@@ -547,281 +379,6 @@ pub fn singleton_field_preamble(
          '{shown_value}'; the field is a singleton — {grammar} — so a sender must not generate \
          more than one field line for it (RFC 9110 §5.3)"
     )
-}
-
-/// Why a value derives from neither alternative of `( token / quoted-string )`.
-///
-/// Data rather than a sentence, and that is the whole reason this extraction was
-/// possible at all. The alternation is one production; the finding is not. Seven
-/// sites read this pair and each says something about its own field — an
-/// `Alt-Svc` parameter, a `Server-Timing` parameter, a `Pragma` directive, a
-/// media type's parameter, a `Keep-Alive` parameter, a `Prefer` preference. Six
-/// of them had also decided, separately and in four different ways, what an
-/// empty value means. So what is shared here is the decision and the two
-/// measurements behind it; **the wording, and the verdict on
-/// [`WordDefect::Empty`], stay at the caller** — which is where each field's own
-/// sentence about it is.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum WordDefect {
-    /// The value is empty, and neither alternative derives the empty string:
-    /// `token = 1*tchar` has a one-character floor, and the shortest
-    /// `quoted-string` is its two DQUOTEs. That is arithmetic on two
-    /// productions and is not a per-field question — but what a field *does*
-    /// about it is, and two rules in this tree tolerate it on the record.
-    Empty,
-    /// An unquoted value holding a character no `tchar` admits. Unquoted by
-    /// elimination rather than by inspection: see [`token_or_quoted_string`].
-    NotToken(char),
-    /// A leading DQUOTE opened something that is not a well-formed
-    /// `quoted-string`, carrying the interior walk's own account of it.
-    NotQuotedString(String),
-}
-
-/// Read one `( token / quoted-string )` and return what it holds.
-///
-/// The alternation is decided by the first octet and by nothing else: RFC 9110
-/// § 5.6.2 makes DQUOTE a delimiter, so no `token` can open with one and a value
-/// that does is trying to be a `quoted-string` and nothing else. A value that
-/// does not open with one is a `token` by elimination, which is why the `tchar`
-/// scan below is the whole of that half.
-///
-/// The content returned is what the field means by the value: a `token` as
-/// written, a `quoted-string` after `quoted-pair` substitution. A caller that
-/// only judges the value may discard it — [`unescape_quoted_string`] opens by
-/// asking [`validate_quoted_string`](crate::helpers::quoted_string::validate_quoted_string), so the two accept exactly the same strings
-/// and the `Err` is the same `Err`.
-///
-/// RFC 7230 named this pair `word`; RFC 9110 kept both halves and dropped the
-/// name, which is why the two productions cited here are the halves and not the
-/// whole. Callers whose own document writes its own name for it —
-/// `server-timing-param-value`, `parameter-value`, RFC 2068's `value` — cite
-/// that at their site.
-///
-// cite(RFC 9110 § 5.6.2): "Delimiters are chosen from the set of US-ASCII visual characters not allowed in a token (DQUOTE and "(),/:;<=>?@[\]{}")."
-// cite(RFC 9110 § 5.6.2): "token = 1*tchar tchar = "!" / "#" / "$" / "%" / "&" / "'" / "*" / "+" / "-" / "." / "^" / "_" / "`" / "|" / "~" / DIGIT / ALPHA"
-// cite(RFC 9110 § 5.6.4): "quoted-string  = DQUOTE *( qdtext / quoted-pair ) DQUOTE"
-pub fn token_or_quoted_string(value: &str) -> Result<std::borrow::Cow<'_, str>, WordDefect> {
-    if value.is_empty() {
-        return Err(WordDefect::Empty);
-    }
-    if value.starts_with('"') {
-        return unescape_quoted_string(value)
-            .map(std::borrow::Cow::Owned)
-            .map_err(WordDefect::NotQuotedString);
-    }
-    match crate::helpers::token::find_invalid_token_char(value) {
-        Some(c) => Err(WordDefect::NotToken(c)),
-        None => Ok(std::borrow::Cow::Borrowed(value)),
-    }
-}
-
-/// One `token [ BWS "=" BWS word ]` pair, parsed.
-///
-/// RFC 7240 writes that production three times — `preference`, `parameter` and
-/// `applied-pref` — and says in prose that the third is the first minus its
-/// parameters, so the pair itself is one thing written once here rather than
-/// per field. `word` is the name RFC 7230 gave `( token / quoted-string )`;
-/// RFC 9110 kept both halves and dropped the name, which is why the pair is
-/// cited by its halves at [`token_or_quoted_string`], where it is read, and not
-/// under a name no document in force writes.
-pub struct TokenBwsWord<'a> {
-    /// The `token`, exactly as written. Case folding is the caller's: whether
-    /// two names are the same string is a question each field answers for
-    /// itself.
-    pub name: &'a str,
-    /// The `word`'s content — `None` when the optional group is absent, and for
-    /// a `quoted-string` the octets after `quoted-pair` substitution rather than
-    /// the quoted form. An empty `Some` can only come from `""`, which some
-    /// fields define as meaning no value; that reading belongs to the field, so
-    /// it is reported here as what was written.
-    pub value: Option<String>,
-    /// Whether whitespace appeared beside the `=`.
-    ///
-    /// The member handed in has already lost the `OWS` its list prints around
-    /// the commas, so anything still adjacent to the `=` is the `BWS` the
-    /// grammar admits only for historical reasons — and that is a statement
-    /// about the sender, which is why this is returned instead of being
-    /// silently absorbed by the trim that has to happen anyway.
-    pub bws: bool,
-}
-
-/// Parse `token [ BWS "=" BWS word ]` from one already-`OWS`-trimmed member.
-///
-/// Callers reading a field through [`combined_field_value_as_written`] hand this
-/// one `char` per octet, and that is the intended input: an `obs-text` octet is
-/// admitted by the `quoted-string` half and by no part of the `token` half, so
-/// only the octets say which side of the alternation it landed on.
-///
-/// The shape itself is cited at the field that has it, not here — a helper
-/// shared by three productions can honestly quote only the halves they agree
-/// on. `token` is transcribed once at [`crate::helpers::token::is_tchar`], and
-/// the `word` half is read by [`token_or_quoted_string`]; what is left here is
-/// the part RFC 7240 owns, which is the optional group and the `BWS`.
-///
-/// cite(RFC 9110 § 5.6.3): "The BWS rule is used where the grammar allows optional whitespace only for historical reasons."
-pub fn parse_token_bws_word(member: &str) -> Result<TokenBwsWord<'_>, String> {
-    // `token` admits neither `=` nor DQUOTE, so nothing can precede the `=` the
-    // production prints except the name -- there is no quoted-string in front of
-    // it for one to hide in, and the first `=` is therefore the delimiter even
-    // when the `word` after it is a quoted-string containing more of them.
-    let (name_written, value_written) = match member.find('=') {
-        Some(i) => (&member[..i], Some(&member[i + 1..])),
-        None => (member, None),
-    };
-    let name = trim_ows(name_written);
-    let value_written_trimmed = value_written.map(trim_ows);
-    let bws = name != name_written || value_written_trimmed != value_written;
-
-    if name.is_empty() {
-        return Err("no token before the \"=\"".to_string());
-    }
-    if let Some(c) = crate::helpers::token::find_invalid_token_char(name) {
-        return Err(format!("token contains {}", describe_char(c)));
-    }
-
-    let value = match value_written_trimmed {
-        None => None,
-        Some(v) => match token_or_quoted_string(v) {
-            Ok(content) => Some(content.into_owned()),
-            // `word` is `token / quoted-string` and `token` is `1*tchar`, so the
-            // optional group cannot close on an empty value: a field that means
-            // to say "no value" writes no `=`, or writes `""`. RFC 7240 grants
-            // no tolerance for the third spelling, so this is a defect here.
-            Err(WordDefect::Empty) => {
-                return Err("nothing after the \"=\", where the grammar has a word".to_string())
-            }
-            Err(WordDefect::NotToken(c)) => {
-                return Err(format!("value contains {}", describe_char(c)))
-            }
-            Err(WordDefect::NotQuotedString(e)) => return Err(e),
-        },
-    };
-
-    Ok(TokenBwsWord { name, value, bws })
-}
-
-/// Validate qvalue syntax: `0`, `1`, `0.5`, `0.123`, `1.000` — up to three
-/// digits after the point, and only zeroes after a leading `1`.
-///
-/// Written as the production reads rather than as a list of the strings it
-/// generates. The enumeration this replaced (`s == "1" || s == "1.0" || ...`)
-/// missed `0.` and `1.`, which `0*3DIGIT` and `0*3("0")` both admit at zero
-/// repetitions — so two conforming values were reported as malformed by every
-/// rule that asks this question.
-///
-/// The leading/trailing trim is a tolerance, not the grammar: a `qvalue` has no
-/// whitespace in it anywhere. Every caller trims before asking, so it changes
-/// no verdict today; it is kept because nothing depends on it being strict and
-/// removing it would be a silent trap for a caller that does not trim.
-pub fn valid_qvalue(s: &str) -> bool {
-    let s = s.trim();
-    // The three-decimal cap and the "1 may only be followed by zeroes" asymmetry are
-    // both in the production; neither is arbitrary.
-    // cite(RFC 9110 § 12.4.2, label: qvalue grammar): "qvalue = ( "0" [ "." 0*3DIGIT ] ) / ( "1" [ "." 0*3("0") ] )"
-    let (rest, digit_ok): (&str, fn(u8) -> bool) = if let Some(rest) = s.strip_prefix('0') {
-        (rest, |b| b.is_ascii_digit())
-    } else if let Some(rest) = s.strip_prefix('1') {
-        (rest, |b| b == b'0')
-    } else {
-        return false;
-    };
-
-    // `[ "." 0*3DIGIT ]` — the fraction is optional, and so is every digit in it.
-    match rest.strip_prefix('.') {
-        None => rest.is_empty(),
-        Some(fraction) => fraction.len() <= 3 && fraction.bytes().all(digit_ok),
-    }
-}
-
-/// Validate an RFC 8187 `ext-value` (e.g. `UTF-8''%e2%82%ac%20rates`).
-/// Returns Ok(()) if the value matches the expected pattern and contains
-/// only allowed characters/percent-escapes, or Err(msg) describing the
-/// problem.
-///
-/// This said RFC 5987, which RFC 8187 obsoletes and moved to Historic. The
-/// pointer is worth correcting and worth not overstating: the two documents'
-/// `ext-value`, `value-chars`, `pct-encoded` and `attr-char` productions are
-/// byte-for-byte identical, so nothing here changed meaning when the reference
-/// did. What 8187 changed is elsewhere -- the ISO-8859-1 requirement is gone,
-/// and it stopped trying to define a generic `parameter` rule.
-///
-/// The `charset` is checked only for being ASCII and quote-free, which is far
-/// looser than `mime-charset`, so that production is deliberately not quoted
-/// below -- it would describe a check this function does not make.
-///
-// cite(RFC 8187 § 3.2.1): "ext-value = charset  "'" [ language ] "'" value-chars"
-pub fn validate_ext_value(val: &str) -> Result<(), String> {
-    // Must contain at least two single quotes separating charset, optional language, and value-chars
-    let first_quote = val
-        .find('\'')
-        .ok_or_else(|| "ext-value missing charset separator".to_string())?;
-    let rest = &val[first_quote + 1..];
-    let second_quote = rest
-        .find('\'')
-        .ok_or_else(|| "ext-value missing language separator".to_string())?
-        + first_quote
-        + 1;
-
-    let charset = &val[..first_quote];
-    if charset.is_empty() {
-        return Err("charset in ext-value must not be empty".into());
-    }
-    // Basic charset sanity: must be ASCII and not contain quote
-    if !charset.is_ascii() || charset.contains('\'') {
-        return Err("invalid charset in ext-value".into());
-    }
-
-    // Language part may be empty; we don't strictly validate language tags here
-    let value_chars = &val[second_quote + 1..];
-    if value_chars.is_empty() {
-        // empty value is allowed
-        return Ok(());
-    }
-
-    let mut i = 0usize;
-    let bytes = value_chars.as_bytes();
-    // cite(RFC 8187 § 3.2.1): "value-chars   = *( pct-encoded / attr-char )"
-    while i < bytes.len() {
-        let b = bytes[i];
-        // This branch is why `%` is absent from the attr-char table below rather
-        // than merely unreachable in it: a `%` here is always the start of an
-        // escape, and the two productions do not overlap.
-        //
-        // cite(RFC 8187 § 3.2.1): "pct-encoded   = "%" HEXDIG HEXDIG"
-        if b == b'%' {
-            // Expect two hex digits
-            if i + 2 >= bytes.len() {
-                return Err("incomplete percent-encoding in ext-value".into());
-            }
-            let hi = bytes[i + 1];
-            let lo = bytes[i + 2];
-            if !((hi as char).is_ascii_hexdigit() && (lo as char).is_ascii_hexdigit()) {
-                return Err("invalid percent-encoding in ext-value".into());
-            }
-            i += 3;
-            continue;
-        }
-        let ch = bytes[i] as char;
-        // The table carried a `'%'` for as long as it has existed, under a comment
-        // claiming it was RFC 5987's attr-char. It was not: both documents spell
-        // this production identically and both exclude `%`, along with `*` and `'`.
-        // It was dead as well as wrong -- the branch above consumes every `%` before
-        // this arm is reached -- so removing it moved no test.
-        //
-        // cite(RFC 8187 § 3.2.1): "attr-char     = ALPHA / DIGIT / "!" / "#" / "$" / "&" / "+" / "-" / "." / "^" / "_" / "`" / "|" / "~" ; token except ( "*" / "'" / "%" )"
-        if ch.is_ascii_alphanumeric()
-            || matches!(
-                ch,
-                '!' | '#' | '$' | '&' | '+' | '-' | '.' | '^' | '_' | '`' | '|' | '~'
-            )
-        {
-            i += 1;
-            continue;
-        }
-        return Err(format!("invalid character '{}' in ext-value", ch));
-    }
-
-    Ok(())
 }
 
 #[cfg(test)]
@@ -914,132 +471,6 @@ mod tests {
         }
     }
 
-    /// Each row is a weight and whether the `qvalue` production admits it. A
-    /// table rather than a run of `assert!` lines, so a failure names the value.
-    #[test]
-    fn test_valid_qvalue() {
-        const CASES: &[(&str, bool)] = &[
-            ("1", true),
-            ("1.0", true),
-            ("1.00", true),
-            ("1.000", true),
-            ("0", true),
-            ("0.5", true),
-            ("0.123", true),
-            ("0.000", true),
-            // `0*3DIGIT` and `0*3("0")` are satisfied by no digits at all, so a
-            // point with nothing after it conforms. The enumeration this replaced
-            // reported both as malformed.
-            ("0.", true),
-            ("1.", true),
-            ("1.0000", false),
-            ("0.1234", false),
-            // Only zeroes may follow a leading 1: the weight is capped at 1.
-            ("1.1", false),
-            ("1.001", false),
-            ("abc", false),
-            ("", false),
-            ("2", false),
-            ("-1", false),
-            ("0.5.5", false),
-            ("00", false),
-            ("0.a", false),
-        ];
-
-        for (value, is_qvalue) in CASES {
-            assert_eq!(valid_qvalue(value), *is_qvalue, "for {value:?}");
-        }
-    }
-
-    /// The alternation is decided by the first octet and by nothing else, and
-    /// each of the three defects is a distinct answer rather than one `Err`:
-    /// the seven callers that used to write this out disagree about `Empty` in
-    /// particular, so it has to arrive as something they can match on.
-    #[test]
-    fn token_or_quoted_string_separates_its_three_defects() {
-        assert_eq!(token_or_quoted_string("abc").unwrap(), "abc");
-        // The content is what the field means by the value, so the DQUOTEs come
-        // off and a `quoted-pair` is substituted.
-        assert_eq!(token_or_quoted_string("\"a\\\"b\"").unwrap(), "a\"b");
-        // `""` is two DQUOTEs and derives from the production; its content is
-        // empty, which is not the same fact as the value being empty.
-        assert_eq!(token_or_quoted_string("\"\"").unwrap(), "");
-
-        assert_eq!(token_or_quoted_string(""), Err(WordDefect::Empty));
-        assert_eq!(
-            token_or_quoted_string("a b"),
-            Err(WordDefect::NotToken(' '))
-        );
-        // A leading DQUOTE commits the value to the `quoted-string` half: RFC
-        // 9110 § 5.6.2 makes DQUOTE a delimiter, so this is not a `token`
-        // holding a bad character, it is a malformed `quoted-string`.
-        assert!(matches!(
-            token_or_quoted_string("\"abc"),
-            Err(WordDefect::NotQuotedString(_))
-        ));
-        assert!(matches!(
-            token_or_quoted_string("\"abc\"x"),
-            Err(WordDefect::NotQuotedString(_))
-        ));
-    }
-
-    /// The reader is handed one `char` per octet by every caller that reads a
-    /// field through [`combined_field_value_as_written`], and the two halves
-    /// answer an `obs-text` octet differently on purpose: `qdtext` admits it and
-    /// `tchar` does not.
-    #[test]
-    fn obs_text_lands_on_the_side_of_the_alternation_the_grammar_puts_it() {
-        let quoted: String = [b'"', 0xE9, b'"'].iter().map(|&b| b as char).collect();
-        assert_eq!(
-            token_or_quoted_string(&quoted).unwrap().chars().count(),
-            1,
-            "an obs-text octet is qdtext and comes back as one octet"
-        );
-        let bare: String = [0xE9u8].iter().map(|&b| b as char).collect();
-        assert_eq!(
-            token_or_quoted_string(&bare),
-            Err(WordDefect::NotToken('\u{E9}'))
-        );
-    }
-
-    /// `%` is not an attr-char. It reaches this function only as the opening of a
-    /// `pct-encoded`, and the branch handling that runs first -- so a `%` that is
-    /// not two hex digits is an incomplete escape, never a literal.
-    #[test]
-    fn percent_is_only_ever_an_escape() {
-        assert!(validate_ext_value("UTF-8''%41").is_ok());
-        assert!(validate_ext_value("UTF-8''a%20b").is_ok());
-        assert!(validate_ext_value("UTF-8''%").is_err());
-        assert!(validate_ext_value("UTF-8''a%b").is_err());
-        assert!(validate_ext_value("UTF-8''100%").is_err());
-    }
-
-    /// The three characters `token` has and `attr-char` does not.
-    #[test]
-    fn attr_char_excludes_star_quote_and_percent() {
-        assert!(validate_ext_value("UTF-8''a*b").is_err());
-        assert!(validate_ext_value("UTF-8''a{b").is_err());
-        assert!(validate_ext_value("UTF-8''ok-name.ext~1").is_ok());
-    }
-
-    #[test]
-    fn test_validate_ext_value() {
-        // Valid ext-values
-        assert!(validate_ext_value("UTF-8''%e2%82%ac%20rates").is_ok());
-        assert!(validate_ext_value("iso-8859-1'en'%A3%20rates").is_ok());
-        assert!(validate_ext_value("UTF-8''simple-ascii").is_ok());
-        assert!(validate_ext_value("UTF-8''").is_ok()); // empty value-chars allowed
-
-        // Invalid: missing quotes
-        assert!(validate_ext_value("UTF-8%e2%82%ac").is_err());
-        // Invalid: incomplete percent
-        assert!(validate_ext_value("UTF-8''%e2%2").is_err());
-        // Invalid: bad hex
-        assert!(validate_ext_value("UTF-8''%ZZ").is_err());
-        // Invalid: invalid attr-char
-        assert!(validate_ext_value("UTF-8''hello@world").is_err());
-    }
-
     /// Wire order, the four labels, and the two ways a section is absent — a
     /// framing that carried no trailers, and an upstream that never answered.
     #[test]
@@ -1090,48 +521,5 @@ mod tests {
                 .collect::<Vec<_>>(),
             vec!["header section", "trailer section"]
         );
-    }
-
-    #[test]
-    fn test_is_prohibited_trailer_field() {
-        // Fields the cited categories name, case-insensitively.
-        assert!(is_prohibited_trailer_field("Connection"));
-        assert!(is_prohibited_trailer_field("connection"));
-        assert!(is_prohibited_trailer_field("keep-alive"));
-        // The categories reach well past the connection-specific ones: framing,
-        // routing, conditionals, response control data.
-        assert!(is_prohibited_trailer_field("Content-Length"));
-        assert!(is_prohibited_trailer_field("host"));
-        assert!(is_prohibited_trailer_field("If-Match"));
-        assert!(is_prohibited_trailer_field("date"));
-        // A Trailer field inside a trailer section announces nothing.
-        assert!(is_prohibited_trailer_field("trailer"));
-        // An extension field is not prohibited by what it is.
-        assert!(!is_prohibited_trailer_field("x-foo"));
-        assert!(!is_prohibited_trailer_field("x-checksum"));
-    }
-
-    #[test]
-    fn test_is_nominated_by_connection() {
-        // Nomination is per-message: this field is disqualified only because this
-        // message's Connection header names it.
-        assert!(is_nominated_by_connection(
-            "X-Special",
-            Some("keep-alive, X-Special")
-        ));
-        // Not nominated if not listed.
-        assert!(!is_nominated_by_connection("X-Special", Some("keep-alive")));
-        // Both sides match case-insensitively.
-        assert!(is_nominated_by_connection(
-            "x-special",
-            Some("KEEP-ALIVE, x-special")
-        ));
-        // No Connection header nominates nothing.
-        assert!(!is_nominated_by_connection("x-special", None));
-        // A substring is not a token.
-        assert!(!is_nominated_by_connection(
-            "upgrade",
-            Some("super-upgrade")
-        ));
     }
 }
