+++ lint-http-rules/src/helpers/mod.rs
@@ -46,6 +46,7 @@ pub mod content_length;
 pub mod content_range;
 pub mod cookie;
 pub mod domain;
+pub mod field_placement;
 pub mod forwarded_node;
 pub mod headers;
 pub mod ipv6;
@@ -56,6 +57,7 @@ pub mod media_type;
 pub mod parameter;
 pub mod product;
 pub mod quoted_string;
+pub mod qvalue;
 pub mod rule_config;
 pub mod shown;
 pub mod status;
@@ -65,3 +67,4 @@ pub mod uri;
 pub mod validator;
 pub mod vary;
 pub mod websocket;
+pub mod word;
