+++ lint-http-rules/src/helpers/parameter.rs
@@ -27,7 +27,7 @@ use crate::helpers::list::split_semicolons_respecting_quotes;
 /// `value` is **as written** — a `quoted-string` still carries its DQUOTEs.
 /// Whether this field's value half is § 5.6.6's `( token / quoted-string )`, a
 /// narrower production, or one literal is the caller's grammar and not this
-/// walk's, and [`token_or_quoted_string`](crate::helpers::headers::token_or_quoted_string) is what reads it where it is the
+/// walk's, and [`token_or_quoted_string`](crate::helpers::word::token_or_quoted_string) is what reads it where it is the
 /// former.
 pub struct Parameter<'a> {
     /// The `parameter-name` as written. Case folding is the caller's — § 5.6.6
@@ -127,6 +127,103 @@ pub fn parameter_of(segment: &str) -> Option<Result<Parameter<'_>, ParameterDefe
     })
 }
 
+// ── `ext-value`, the other form a parameter value takes ──────────────
+//
+// RFC 8187's `filename*=UTF-8''...`. Not a separate question and so not a
+// separate module: it is what a parameter value may be when the parameter name
+// ends in `*`, and every reader of it is already reading parameters.
+
+/// Validate an RFC 8187 `ext-value` (e.g. `UTF-8''%e2%82%ac%20rates`).
+/// Returns Ok(()) if the value matches the expected pattern and contains
+/// only allowed characters/percent-escapes, or Err(msg) describing the
+/// problem.
+///
+/// This said RFC 5987, which RFC 8187 obsoletes and moved to Historic. The
+/// pointer is worth correcting and worth not overstating: the two documents'
+/// `ext-value`, `value-chars`, `pct-encoded` and `attr-char` productions are
+/// byte-for-byte identical, so nothing here changed meaning when the reference
+/// did. What 8187 changed is elsewhere -- the ISO-8859-1 requirement is gone,
+/// and it stopped trying to define a generic `parameter` rule.
+///
+/// The `charset` is checked only for being ASCII and quote-free, which is far
+/// looser than `mime-charset`, so that production is deliberately not quoted
+/// below -- it would describe a check this function does not make.
+///
+// cite(RFC 8187 § 3.2.1): "ext-value = charset  "'" [ language ] "'" value-chars"
+pub fn validate_ext_value(val: &str) -> Result<(), String> {
+    // Must contain at least two single quotes separating charset, optional language, and value-chars
+    let first_quote = val
+        .find('\'')
+        .ok_or_else(|| "ext-value missing charset separator".to_string())?;
+    let rest = &val[first_quote + 1..];
+    let second_quote = rest
+        .find('\'')
+        .ok_or_else(|| "ext-value missing language separator".to_string())?
+        + first_quote
+        + 1;
+
+    let charset = &val[..first_quote];
+    if charset.is_empty() {
+        return Err("charset in ext-value must not be empty".into());
+    }
+    // Basic charset sanity: must be ASCII and not contain quote
+    if !charset.is_ascii() || charset.contains('\'') {
+        return Err("invalid charset in ext-value".into());
+    }
+
+    // Language part may be empty; we don't strictly validate language tags here
+    let value_chars = &val[second_quote + 1..];
+    if value_chars.is_empty() {
+        // empty value is allowed
+        return Ok(());
+    }
+
+    let mut i = 0usize;
+    let bytes = value_chars.as_bytes();
+    // cite(RFC 8187 § 3.2.1): "value-chars   = *( pct-encoded / attr-char )"
+    while i < bytes.len() {
+        let b = bytes[i];
+        // This branch is why `%` is absent from the attr-char table below rather
+        // than merely unreachable in it: a `%` here is always the start of an
+        // escape, and the two productions do not overlap.
+        //
+        // cite(RFC 8187 § 3.2.1): "pct-encoded   = "%" HEXDIG HEXDIG"
+        if b == b'%' {
+            // Expect two hex digits
+            if i + 2 >= bytes.len() {
+                return Err("incomplete percent-encoding in ext-value".into());
+            }
+            let hi = bytes[i + 1];
+            let lo = bytes[i + 2];
+            if !((hi as char).is_ascii_hexdigit() && (lo as char).is_ascii_hexdigit()) {
+                return Err("invalid percent-encoding in ext-value".into());
+            }
+            i += 3;
+            continue;
+        }
+        let ch = bytes[i] as char;
+        // The table carried a `'%'` for as long as it has existed, under a comment
+        // claiming it was RFC 5987's attr-char. It was not: both documents spell
+        // this production identically and both exclude `%`, along with `*` and `'`.
+        // It was dead as well as wrong -- the branch above consumes every `%` before
+        // this arm is reached -- so removing it moved no test.
+        //
+        // cite(RFC 8187 § 3.2.1): "attr-char     = ALPHA / DIGIT / "!" / "#" / "$" / "&" / "+" / "-" / "." / "^" / "_" / "`" / "|" / "~" ; token except ( "*" / "'" / "%" )"
+        if ch.is_ascii_alphanumeric()
+            || matches!(
+                ch,
+                '!' | '#' | '$' | '&' | '+' | '-' | '.' | '^' | '_' | '`' | '|' | '~'
+            )
+        {
+            i += 1;
+            continue;
+        }
+        return Err(format!("invalid character '{}' in ext-value", ch));
+    }
+
+    Ok(())
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -186,5 +283,43 @@ mod tests {
         // flag reads exactly what it read before the flag existed.
         let p = parameters("a = b").next().unwrap().ok().unwrap();
         assert_eq!((p.name, p.value), ("a", "b"));
+    }
+
+    /// `%` is not an attr-char. It reaches this function only as the opening of a
+    /// `pct-encoded`, and the branch handling that runs first -- so a `%` that is
+    /// not two hex digits is an incomplete escape, never a literal.
+    #[test]
+    fn percent_is_only_ever_an_escape() {
+        assert!(validate_ext_value("UTF-8''%41").is_ok());
+        assert!(validate_ext_value("UTF-8''a%20b").is_ok());
+        assert!(validate_ext_value("UTF-8''%").is_err());
+        assert!(validate_ext_value("UTF-8''a%b").is_err());
+        assert!(validate_ext_value("UTF-8''100%").is_err());
+    }
+
+    /// The three characters `token` has and `attr-char` does not.
+    #[test]
+    fn attr_char_excludes_star_quote_and_percent() {
+        assert!(validate_ext_value("UTF-8''a*b").is_err());
+        assert!(validate_ext_value("UTF-8''a{b").is_err());
+        assert!(validate_ext_value("UTF-8''ok-name.ext~1").is_ok());
+    }
+
+    #[test]
+    fn test_validate_ext_value() {
+        // Valid ext-values
+        assert!(validate_ext_value("UTF-8''%e2%82%ac%20rates").is_ok());
+        assert!(validate_ext_value("iso-8859-1'en'%A3%20rates").is_ok());
+        assert!(validate_ext_value("UTF-8''simple-ascii").is_ok());
+        assert!(validate_ext_value("UTF-8''").is_ok()); // empty value-chars allowed
+
+        // Invalid: missing quotes
+        assert!(validate_ext_value("UTF-8%e2%82%ac").is_err());
+        // Invalid: incomplete percent
+        assert!(validate_ext_value("UTF-8''%e2%2").is_err());
+        // Invalid: bad hex
+        assert!(validate_ext_value("UTF-8''%ZZ").is_err());
+        // Invalid: invalid attr-char
+        assert!(validate_ext_value("UTF-8''hello@world").is_err());
     }
 }
