+++ lint-http-rules/src/rules/accept_encoding_parameter_valid.rs
@@ -191,7 +191,7 @@ impl Rule for AcceptEncodingParameterValid {
                                 };
 
                                 // cite(RFC 9110 § 12.4.2): "qvalue = ( "0" [ "." 0*3DIGIT ] ) / ( "1" [ "." 0*3("0") ] )"
-                                if !crate::helpers::headers::valid_qvalue(v) {
+                                if !crate::helpers::qvalue::valid_qvalue(v) {
                                     return Some(self.cited(
                                         &RFC_9110_12_4_2,
                                         ctx.severity,
