+++ lint-http-rules/src/rules/alt_svc_header_syntax.rs
@@ -2,14 +2,13 @@
 //
 // SPDX-License-Identifier: ISC
 
-use crate::helpers::headers::{
-    combined_field_value_as_written, token_or_quoted_string, trim_ows, WordDefect,
-};
+use crate::helpers::headers::{combined_field_value_as_written, trim_ows};
 use crate::helpers::list::{
     list_members_as_written, quoting_is_balanced, split_semicolons_respecting_quotes,
 };
 use crate::helpers::quoted_string::unescape_quoted_string;
 use crate::helpers::shown::{describe_char, shown_in_finding};
+use crate::helpers::word::{token_or_quoted_string, WordDefect};
 use crate::lint::Violation;
 use crate::rules::Rule;
