+++ lint-http-rules/src/rules/link_header_valid.rs
@@ -2,10 +2,11 @@
 //
 // SPDX-License-Identifier: ISC
 
-use crate::helpers::headers::{combined_field_value_as_written, parse_token_bws_word, trim_ows};
+use crate::helpers::headers::{combined_field_value_as_written, trim_ows};
 use crate::helpers::list::split_semicolons_respecting_quotes;
 use crate::helpers::shown::{describe_char, shown_in_finding};
 use crate::helpers::uri::{find_non_uri_char, validate_scheme_name};
+use crate::helpers::word::parse_token_bws_word;
 use crate::lint::Violation;
 use crate::rules::Rule;
