+++ lint-http-rules/src/rules/prefer_header_and_preference_applied.rs
@@ -2,9 +2,10 @@
 //
 // SPDX-License-Identifier: ISC
 
-use crate::helpers::headers::{combined_field_value_as_written, parse_token_bws_word};
+use crate::helpers::headers::combined_field_value_as_written;
 use crate::helpers::list::list_members_as_written;
 use crate::helpers::shown::shown_in_finding;
+use crate::helpers::word::parse_token_bws_word;
 use crate::lint::Violation;
 use crate::rules::Rule;
