+++ lint-http-rules/src/rules/preference_applied_header_valid.rs
@@ -2,11 +2,12 @@
 //
 // SPDX-License-Identifier: ISC
 
-use crate::helpers::headers::{combined_field_value_as_written, parse_token_bws_word};
+use crate::helpers::headers::combined_field_value_as_written;
 use crate::helpers::list::{
     list_members_as_written, quoting_is_balanced, split_semicolons_respecting_quotes,
 };
 use crate::helpers::shown::shown_in_finding;
+use crate::helpers::word::parse_token_bws_word;
 use crate::lint::Violation;
 use crate::rules::Rule;
 use std::collections::HashMap;
